+++ core/native/src/Instant.kt
@@ -274,8 +274,8 @@ actual fun Instant.plus(period: DateTimePeriod, timeZone: TimeZone): Instant = t
                 plus(hours.toLong() * SECONDS_PER_HOUR, 0).check(timeZone) else this }
             .run { if (minutes != 0)
                 plus(minutes.toLong() * SECONDS_PER_MINUTE, 0).check(timeZone) else this }
-            .run { if (seconds != 0L) plus(seconds, 0).check(timeZone) else this }
-            .run { if (nanoseconds != 0L) plus(0, nanoseconds).check(timeZone) else this }
+            .run { if (seconds != 0) plus(seconds.toLong(), 0).check(timeZone) else this }
+            .run { if (nanoseconds != 0) plus(0, nanoseconds.toLong()).check(timeZone) else this }
     }.check(timeZone)
 } catch (e: ArithmeticException) {
     throw DateTimeArithmeticException("Arithmetic overflow when adding CalendarPeriod to an Instant", e)
@@ -328,7 +328,7 @@ actual fun Instant.periodUntil(other: Instant, timeZone: TimeZone): DateTimePeri
     val time = thisLdt.until(otherLdt, DateTimeUnit.NANOSECOND).nanoseconds // |otherLdt - thisLdt| < 24h
 
     time.toComponents { hours, minutes, seconds, nanoseconds ->
-        return DateTimePeriod((months / 12), (months % 12), days, hours, minutes, seconds.toLong(), nanoseconds.toLong())
+        return DateTimePeriod((months / 12), (months % 12), days, hours, minutes, seconds, nanoseconds.toLong())
     }
 }
